compose message service: building and shared to extension. code clean needed
