+++ FlowCrypt/Functionality/Mail Provider/UsersMailSession Provider/Gmail+UsersMailSessionProvider.swift
@@ -15,4 +15,3 @@ extension GmailService: UsersMailSessionProvider {
         userService.renewSession()
     }
 }
-
