+++ FlowCrypt/Functionality/Mail Provider/UsersMailSession Provider/Imap+UserMailSessionProvider.swift
@@ -16,4 +16,3 @@ extension Imap: UsersMailSessionProvider {
         Promise(self.setupSession())
     }
 }
-
